refactor to remove csv
