+++ lib/sycamore/sycamore/writers/duckdb_writer.py
@@ -1,50 +1,150 @@
-from typing import Optional
+from dataclasses import dataclass, asdict, is_dataclass
+from typing import Optional, Any
+from typing_extensions import TypeGuard
 
 from sycamore.data.document import Document
-from sycamore.plan_nodes import Node, Write
-from sycamore.transforms.map import MapBatch
-import glob
-import duckdb
+from sycamore.writers.base import BaseDBWriter
 import os
+import pyarrow as pa
+import duckdb
 
 
-class DuckDB_Writer(MapBatch, Write):
+@dataclass
+class DuckDBClientParams(BaseDBWriter.ClientParams):
+    pass
 
-    def __init__(
-        self,
-        plan: Node,
-        csv_location: str,
-        db_url: str,
-        table_name: Optional[str] = None,
-        csv_directory_location: Optional[str] = None,
-        **kwargs,
+
+@dataclass
+class DuckDBTargetParams(BaseDBWriter.TargetParams):
+    db_url: Optional[str] = ":default:"
+    table_name: Optional[str] = "default_table"
+    persist: Optional[bool] = True
+
+    def compatible_with(self, other: BaseDBWriter.TargetParams) -> bool:
+        if not isinstance(other, DuckDBTargetParams):
+            return False
+        if self.db_url != other.db_url:
+            return False
+        if self.table_name != other.table_name:
+            return False
+        return True
+
+
+class DuckDBClient(BaseDBWriter.Client):
+    def __init__(self, client_params: DuckDBClientParams):
+        pass
+
+    @classmethod
+    def from_client_params(cls, params: BaseDBWriter.ClientParams) -> "DuckDBClient":
+        assert isinstance(params, DuckDBClientParams)
+        return DuckDBClient(params)
+
+    def write_many_records(
+        self, records: list[BaseDBWriter.Record], target_params: BaseDBWriter.TargetParams, batch_size: int = 1000
     ):
-        super().__init__(plan, f=self.write_docs, **kwargs)
-        self._filter = filter
-        self.csv_location = csv_location
-        self.csv_directory_location = csv_directory_location
-        self.db_url = db_url
-        self.table_name = table_name
+        N = 1000 ^ 2  # Bit lesser than 1 MB
+        dict_params = (
+            asdict(target_params)
+            if is_dataclass(target_params)
+            else (target_params.__dict__ if hasattr(target_params, "__dict__") else {})
+        )
+        assert _narrow_list_of_doc_records(records), f"Found a bad record in {records}"
+        assert isinstance(target_params, DuckDBTargetParams)
 
-    def write_docs(self, docs: list[Document]) -> list[Document]:
-        # Check if files are written (to gracefully handle tests to only check execution)
-        sql_location = os.path.join(self.csv_location, "*.csv")
-        self.table_name = self.table_name if not None else "data"
-        if bool(glob.glob(sql_location)):
-            client = duckdb.connect(self.db_url)
-            client.sql(f"CREATE TABLE {self.table_name} AS SELECT * FROM read_csv('{sql_location}')")
-            # Flush out the csv files if not persisted
-            if not self.csv_directory_location:
-                try:
-                    for root, _, files in os.walk(self.csv_location):
-                        for file in files:
-                            file_path = os.path.join(root, file)
-                            try:
-                                os.unlink(file_path)
-                            except Exception as e:
-                                print(f"Error deleting {file_path}: {e}")
-                except Exception as e:
-                    print(f"Error deleting files in {self.csv_location}: {e}")
-        else:
-            print(f"No files in directory matching the pattern in {sql_location}")
-        return docs
+        headers = ["uuid", "embeddings", "properties", "text_representation", "bbox", "shingles", "type"]
+        creation = True
+        schema = pa.schema(
+            [
+                ("uuid", pa.string()),
+                ("embeddings", pa.list_(pa.float32())),
+                ("properties", pa.struct([])),  # Adjust based on your properties structure
+                ("text_representation", pa.string()),
+                ("bbox", pa.string()),
+                ("shingles", pa.list_(pa.int64())),
+                ("type", pa.string()),
+            ]
+        )
+
+        def write_batch(batch_data: dict):
+            # If the file doesn't exist, write headers first
+            pa_table = pa.Table.from_pydict(batch_data)
+            client = duckdb.connect(str(dict_params.get("db_url")))
+            nonlocal creation
+            if creation:
+                client.sql(f"CREATE TABLE {dict_params.get('table_name')} AS SELECT * FROM pa_table")
+                creation = False
+            else:
+                client.sql(f"INSERT INTO {dict_params.get('table_name')} SELECT * FROM pa_table")
+            for key in batch_data:
+                batch_data[key].clear()
+
+        batch_data: dict[str, list[Any]] = {key: [] for key in headers}
+
+        for r in records:
+            # Append the new data to the batch
+            batch_data["uuid"].append(r.uuid)
+            batch_data["embeddings"].append(r.embeddings)
+            batch_data["properties"].append(r.properties)
+            batch_data["text_representation"].append(r.text_representation)
+            batch_data["bbox"].append(r.bbox)
+            batch_data["shingles"].append(r.shingles)
+            batch_data["type"].append(r.type)
+
+            # If we've reached the batch size, write to the database
+            if batch_data.__sizeof__() >= N:
+                write_batch(batch_data)
+        # Write any remaining records
+        if len(batch_data["uuid"]) > 0:
+            write_batch(batch_data)
+        # Flush out the database if not persisted
+        if not target_params.persist:
+            try:
+                os.unlink(str(dict_params.get("db_url")))
+            except Exception as e:
+                print(f"Error deleting {dict_params.get('db_url')}: {e}")
+
+    def create_target_idempotent(self, target_params: BaseDBWriter.TargetParams):
+        assert isinstance(target_params, DuckDBTargetParams)
+
+    def get_existing_target_params(self, target_params: BaseDBWriter.TargetParams) -> "DuckDBTargetParams":
+        assert isinstance(target_params, DuckDBTargetParams)
+        return DuckDBTargetParams(db_url=target_params.db_url, table_name=target_params.table_name)
+
+
+@dataclass
+class DuckDBDocumentRecord(BaseDBWriter.Record):
+    uuid: str
+    embeddings: Optional[list[float]] = None
+    properties: Optional[dict[str, Any]] = None
+    text_representation: Optional[str] = None
+    bbox: Optional[tuple[float, float, float, float]] = None
+    shingles: Optional[list[int]] = None
+    type: Optional[str] = None
+
+    @classmethod
+    def from_doc(cls, document: Document, target_params: BaseDBWriter.TargetParams) -> "DuckDBDocumentRecord":
+        assert isinstance(target_params, DuckDBTargetParams)
+        uuid = document.doc_id
+        if uuid is None:
+            raise ValueError(f"Cannot write documents without a doc_id. Found {document}")
+        embedding = document.embedding
+        return DuckDBDocumentRecord(
+            uuid=uuid,
+            properties=document.properties,
+            type=document.type,
+            text_representation=document.text_representation,
+            bbox=document.bbox.coordinates if document.bbox else None,
+            shingles=document.shingles,
+            embeddings=embedding,
+        )
+
+
+def _narrow_list_of_doc_records(records: list[BaseDBWriter.Record]) -> TypeGuard[list[DuckDBDocumentRecord]]:
+    return all(isinstance(r, DuckDBDocumentRecord) for r in records)
+
+
+class DuckDBWriter(BaseDBWriter):
+    Client = DuckDBClient
+    Record = DuckDBDocumentRecord
+    ClientParams = DuckDBClientParams
+    TargetParams = DuckDBTargetParams
